Clean up the Python tests
